LRJS R.org refacor main-concepts: bind method used in event handler
See this article:
@link https://reactjs.org/docs/handling-events.html

Methods which are used in the event handlers (which somehow are like public
methods) have to bind. Otherwise they don't work correctly and the `this` will
be undefined.
It's possible to use experimental public class field syntax
(`methodName = () => { return 'something'; }`), instead binding the method in
constructor. But as a warning you should know this is an experimental syntax
which means it needs to be enabled in the library or framework. However this is
enabled by default if setup a project with `npx create-react-app my-app`, but
using it in another ready project needs to make sure the feature is enabled.
Another issue is, if the feature disables later, we will run into error, so it
seems using constructor binding is the safest approach to use.